+++ src/components/EditionDetail/Edition.tsx
@@ -16,7 +16,6 @@ import {
   TemplateContent,
   TemplateContentPrimary,
   TemplateContentTop,
-  TemplateFooter,
   Text,
   Toggle,
 } from "@nypl/design-system-react-components";
@@ -199,7 +198,6 @@ const Edition: React.FC<{ editionResult: EditionResult; backUrl?: string }> = (
           </SimpleGrid>
         </TemplateContentPrimary>
       </TemplateContent>
-      <TemplateFooter />
     </Template>
   );
 };