+++ src/components/Landing/Landing.tsx
@@ -10,7 +10,6 @@ import {
   TemplateBreakout,
   TemplateContent,
   TemplateContentPrimary,
-  TemplateFooter,
   TemplateHeader,
   useNYPLBreakpoints,
 } from "@nypl/design-system-react-components";
@@ -24,15 +23,18 @@ const LandingPage: React.FC<{ collections?: Opds2Feed }> = ({
   collections,
 }) => {
   const subHeader = (
-    <Box>
+    <Box
+      sx={{
+        a: {
+          color: "ui.link.primary",
+        },
+      }}
+    >
       <span>
         Find millions of digital books for research from multiple sources
         world-wide--all free to read, download, and keep. No library card
         required. This is an early beta test, so we want your feedback!{" "}
-        <Link href="/about" color="ui.link.primary">
-          Read more about the project
-        </Link>
-        .
+        <Link href="/about">Read more about the project</Link>.
       </span>
       <Box marginTop="s">
         <SearchForm />
@@ -80,6 +82,7 @@ const LandingPage: React.FC<{ collections?: Opds2Feed }> = ({
           <Hero
             backgroundColor="#E9E9E9"
             backgroundImageSrc={backgroundImageSrc}
+            foregroundColor="black"
             heroType="primary"
             heading={
               <Heading level="one" id="primary-hero">
@@ -87,14 +90,6 @@ const LandingPage: React.FC<{ collections?: Opds2Feed }> = ({
               </Heading>
             }
             subHeaderText={subHeader}
-            sx={{
-              "> div": {
-                color: "ui.black",
-              },
-              "div a": {
-                color: "ui.link.primary",
-              },
-            }}
           />
         </TemplateHeader>
       </TemplateBreakout>
@@ -106,7 +101,6 @@ const LandingPage: React.FC<{ collections?: Opds2Feed }> = ({
           </Box>
         </TemplateContentPrimary>
       </TemplateContent>
-      <TemplateFooter />
     </Template>
   );
 };